# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HRPF API",
    "version": "0.4.0",
    "description": "Backend foundations, managed public content and atomic publication controls. Fixed NGO page copy is bundled in the frontend and needs no publication approval. Operational member/case review workflows remain later milestones. Public reads are no-store and withdrawals take effect immediately. Browser clients use the frontend same-origin /api rewrite and credentials. All mutating auth/admin requests require the exact configured Origin and X-CSRF-Token. Production cookies use Secure, HttpOnly, SameSite=Lax, Path=/, no Domain; development names omit __Host-."
  },
  "paths": {
...
    "/api/settings/public": {
      "get": {
        "summary": "Read whitelisted public settings with field-level secret exclusion",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "identity": {
                          "type": "object",
                          "properties": {
                            "name": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 1000
                            },
                            "shortName": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 1000
                            },
                            "type": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 1000
                            },
                            "website": {
                              "type": "string",
                              "format": "uri"
                            },
                            "visionLine": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 1000
                            }
                          },
                          "required": [
                            "name",
                            "shortName",
                            "type",
                            "website",
                            "visionLine"
                          ]
                        },
                        "contact": {
                          "type": "object",
                          "properties": {
                            "address": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 1000
                            },
                            "postalCode": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 1000
                            },
                            "phone": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 1000
                            },
                            "landline": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 1000
                            },
                            "emails": {
                              "maxItems": 5,
                              "type": "array",
                              "items": {
                                "type": "string",
                                "maxLength": 254,
                                "format": "email",
                                "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                              }
                            }
                          },
                          "required": [
                            "address",
                            "postalCode",
                            "phone",
                            "landline",
                            "emails"
                          ]
                        },
                        "socialLinks": {
                          "type": "object",
                          "properties": {
                            "facebook": {
                              "type": "string",
                              "format": "uri"
                            },
                            "youtube": {
                              "type": "string",
                              "format": "uri"
                            },
                            "x": {
                              "type": "string",
                              "format": "uri"
                            },
                            "linkedin": {
                              "type": "string",
                              "format": "uri"
                            },
                            "tiktok": {
                              "type": "string",
                              "format": "uri"
                            },
                            "threadsHandle": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 1000
                            }
                          }
                        },
                        "donations": {
                          "type": "object",
                          "properties": {
                            "accountTitle": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 1000
                            },
                            "bank": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 1000
                            },
                            "branch": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 1000
                            },
                            "accountNumber": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 1000
                            },
                            "iban": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 1000
                            },
                            "jazzCash": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 1000
                            }
                          },
                          "required": [
                            "accountTitle",
                            "bank",
                            "branch",
                            "accountNumber",
                            "iban",
                            "jazzCash"
                          ]
                        }
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Redacted validation, authorization, conflict, limit or dependency error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/blogs/{slug}": {
      "get": {
        "summary": "Read a reviewed published article; drafts and future releases return 404",
        "security": [],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
              "maxLength": 100
            }
          },
          {
            "name": "locale",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "ur"
              ],
              "default": "en"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 48,
              "default": 12
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 80
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "media-coverage",
                "in-action"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "title": {
                          "type": "string"
                        },
                        "blocks": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "type": {
                                "type": "string",
                                "enum": [
                                  "paragraph",
                                  "heading",
                                  "list"
                                ]
                              },
                              "text": {
                                "type": "string"
                              },
                              "items": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              }
                            },
                            "required": [
                              "type"
                            ]
                          }
                        },
                        "slug": {
                          "type": "string"
                        },
                        "excerpt": {
                          "type": "string"
                        },
                        "publishedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "title",
                        "blocks",
                        "slug",
                        "excerpt",
                        "publishedAt"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Redacted validation, authorization, conflict, limit or dependency error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/board": {
      "get": {
        "summary": "Read explicitly projected published records; no provider URLs or private source fields",
        "security": [],
        "parameters": [
          {
            "name": "locale",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "ur"
              ],
              "default": "en"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 48,
              "default": 12
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 80
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "media-coverage",
                "in-action"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "designation": {
                            "type": "string"
                          },
                          "slotLabel": {
                            "type": "string"
                          },
                          "rank": {
                            "type": "number"
                          },
                          "bio": {
                            "type": "string"
                          },
                          "photo": {
                            "type": [
                              "string",
                              "null"
                            ]
                          }
                        },
                        "required": [
                          "name",
                          "slug",
                          "designation",
                          "rank",
                          "bio",
                          "photo"
                        ]
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Redacted validation, authorization, conflict, limit or dependency error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/blog-categories": {
      "get": {
        "summary": "Read explicitly projected published records; no provider URLs or private source fields",
        "security": [],
        "parameters": [
          {
            "name": "locale",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "ur"
              ],
              "default": "en"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 48,
              "default": 12
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 80
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "media-coverage",
                "in-action"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "slug": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "slug",
                          "name"
                        ]
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Redacted validation, authorization, conflict, limit or dependency error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/public-assets/{id}": {
      "get": {
        "summary": "Recheck released entity and clean bound file before streaming; never expose restricted originals",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "No-store reviewed file; PDF attachment or inline raster image. Report downloads increment after successful GET, never HEAD.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "default": {
            "description": "Redacted validation, authorization, conflict, limit or dependency error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/{id}/download": {
      "get": {
        "summary": "Recheck released entity and clean bound file before streaming; never expose restricted originals",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "No-store reviewed file; PDF attachment or inline raster image. Report downloads increment after successful GET, never HEAD.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "default": {
            "description": "Redacted validation, authorization, conflict, limit or dependency error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "fields": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    },
    "securitySchemes": {
      "accessCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-hrpf-access"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-hrpf-refresh"
      },
      "csrfCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-hrpf-csrf"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "formTicketHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Form-Ticket"
      }
    }
  }
}